[chore] Working model completed
